[refactor-#222] UserStudyState ViewModel으로 통합 및 매핑 수정
